Move the `mike` version sorting CLI tool
The `mike` version sorting CLI tool is now in
`src/frequenz/repo/config/cli/version/mike/sort.py` instead of in
`src/frequenz/repo/config/mkdocs/sort_versions.py`.

It can be invoked with `python -m
frequenz.repo.config.cli.version.mike.sort`.
